Major changes in client and server
Refactored the client to make connecting to the server a separate
function call, to allow us to actually write unittests.

Switched the server to using a serverController which then spawns the
server. Allows us to use Actions on the server side as well. (Although
the actual actions are not yet hooked up)

Added a displayerrors GUI object to make use of the refactored Client
code to display an error when unable to connect, rather than just
exiting.