Added ability to set a custom player name
In order to accomplish this, I was forced to change how
clientControllerStates hold info, which was the majority of the change.

Outside of that, the server now expects a client to send them a name
before anything else after connecting. If you send an empty string, you
get the default name.

createserverwindow does not currently have a player name space, so it
always sends an empty string

Client is now able to send an arbitrary protobuf to the server.
We should consider moving the functionality in the serverControllerState
into the server as well, it makes more sense to me.